Cleanup code and move Diff to dedicated submodule
